pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cortex-mem"
dynamic = ["version"]
description = "Always-On Memory Service with Progressive Disclosure (L0/L1/L2) and Weighted Retrieval"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "dhawalc", email = "dhawal@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "chromadb>=0.4.0",
    "aiosqlite>=0.19.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]

[project.scripts]
cortex-mem = "cortex_mem.cli:main"

[project.entry-points."openclaw.plugins"]
cortex-mem = "cortex_mem.openclaw_plugin:MemoryPlugin"

[project.urls]
Homepage = "https://github.com/dhawalc/cortex-mem"
Repository = "https://github.com/dhawalc/cortex-mem"
Issues = "https://github.com/dhawalc/cortex-mem/issues"

[tool.setuptools.dynamic]
version = {attr = "cortex_mem.__version__.__version__"}

[tool.setuptools.packages.find]
include = ["cortex_mem*", "service*", "cortex*"]
exclude = ["tests*", "scripts*", "snapshots*", "index*"]
